requirement.txt
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2017.11.5
chardet==3.0.4
class-registry==2.1.2
cobble==0.1.3
colorama==0.4.1
confusable-homoglyphs==3.2.0
cycler==0.10.0
deap==1.2.2
decorator==4.4.0
Django==2.1.7
django-authtools==1.6.0
django-bootstrap==0.2.4
django-bootstrap3==11.0.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-filter==0.14.0
django-filters==0.2.1
django-filters-mixin==0.1.4
django-mongodb-engine==0.6.0
django-nonrel-enuff==0.4
django-registration==2.4.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
djangotoolbox==1.8.0
docx2txt==0.7
easy-thumbnails==2.5
elasticsearch==5.5.1
elasticsearch-dsl==5.3.0
et-xmlfile==1.0.1
filters==1.3.2
filters-django==1.0.5
get==2018.11.19
h5py==2.9.0
haversine==2.1.1
idna==2.5
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
mammoth==1.4.9
matplotlib==3.0.2
metakernel==0.20.14
mongoengine==0.15.0
networkx==2.3
nltk==3.4
numpy==1.16.1
openpyxl==2.5.14
pandas==0.24.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
post==2018.11.20
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
public==2018.11.20
Pygments==2.3.1
pymongo==3.6.0
pyparsing==2.3.1
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2018.7
PyYAML==5.1
pyzmq==18.0.1
query-string==2018.11.20
regex==2019.2.7
request==2018.11.20
requests==2.21.0
scikit-learn==0.20.2
scilab-kernel==0.9.5
scipy==1.2.0
seaborn==0.9.0
selenium==3.141.0
simplejsonrpc==1.2.1
simplejsonseq==0.2.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
soupsieve==1.7.3
sqlparse==0.2.4
str==0.1
tabula==1.0.5
tika==1.19
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.14.1
xlrd==1.2.0

